pyproject.toml
[project]
name = "fast-zero"
version = "0.1.0"
description = ""
authors = [
    { name = "KrodrigoDev", email = "kauanrodrigoo25@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pydantic-settings (>=2.10.0,<3.0.0)",
    "alembic (>=1.16.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.11"
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
taskipy = "^1.14.1"

[tool.ruff]
line-length = 79 # limitando a quantidade de caracteres por linha
extend-exclude = ['migrations'] # pasta que ele não vai olhar

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']


[tool.ruff.format]
preview = true
quote-style = 'single'


[tool.pytest.ini_options]
pythonpath = '.'
addopts = '-p no:warnings'


[tool.taskipy.tasks]
lint = 'ruff check'
format = 'ruff format'
run = 'fastapi dev fast_zero/app.py'
test = 'pytest -s -x --cov=fast_zero -vv'
pre_format = 'ruff check --fix'
pre_test = 'task lint'
post_test = 'coverage html'
servers_run = 'Get-Process | Where-Object {$_.Path -like "*fastapi*"}'
kill_server = 'Stop-Process -Name python'

pre_migration = "alembic revision --autogenerate -m 'upgrade table'"
post_migration = 'alembic upgrade head'